# Optimize IntSet.Bin