… containers-tests/tests/intmap-properties.hs
… containers-tests/tests/intset-properties.hs