@treeowl 2024-04-06
This isn't related to your changes, but it looks like the `IntSet` validity test doesn't ensure that each `Tip` contains at least one element.