@treeowl 2024-06-04
Yes, I'm so sorry about the delay. Problems continue on my end, but that's no reason to hold you up. Merged! Will you try the `IntSet` now?