@treeowl 2024-06-04
No, I'm just incredibly tired from what's been going on at home.

On Tue, Jun 4, 2024, 9:43 AM Soumik Sarkar ***@***.***> wrote:

> Thank you!
>
> Will you try the IntSet now?
>
> This was the IntSet PR. Are you thinking of something else?
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/haskell/containers/pull/998#issuecomment-2147573834>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAOOF7IFP3VXBJOJRKYWDLLZFXAAHAVCNFSM6AAAAABF2HCEOWVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDCNBXGU3TGOBTGQ>
> .
> You are receiving this because you modified the open/close state.Message
> ID: ***@***.***>
>